TST: Move abort test into a separate file.
Then call it with subprocess so that the SIGINT does not kill the pytest
thread.
